Added a global type name registration scheme for Routers, Policies, and Transports.
This now pushes the responsibility for deciding which modules should be supported to the application, which has to do that anyway if they want those modules to be compiled in.  Now, the same syntax that includes the module in the build (the underscore include) will also register the type with the global name repository.  This just allows the deserialization of that type from JSON.
