--- beware_of_leopard.go
@@ -1,9 +1,0 @@
-package ratnet
-
-// Dummy function to make 'go get' happy.
-func Dummy() {
-	// As a method of avoiding circular dependencies,
-	// we're keeping the base package devoid of any actual functionality
-	//
-	// We Apologize for the Inconvenience
-}
