+++ nodes/ram/ramnode.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/awgh/bencrypt/bc"
 	"github.com/awgh/ratnet/api"
+	"github.com/awgh/ratnet/router"
 )
 
 type outboxMsg struct {
@@ -61,7 +62,7 @@ func New(contentKey, routingKey bc.KeyPair) *Node {
 	node.err = make(chan api.Msg)
 
 	// setup default router
-	node.router = api.NewDefaultRouter()
+	node.router = router.NewDefaultRouter()
 
 	return node
 }
